# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedTerms API - Medical Terms Dictionary and Flashcard App
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TermDifficulty = typeof TermDifficulty[keyof typeof TermDifficulty];


export const TermDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface Term {
  id: number;
  term: string;
  /** @nullable */
  pronunciation?: string | null;
  definition: string;
  /** @nullable */
  example?: string | null;
  category: string;
  difficulty: TermDifficulty;
  isFavorite: boolean;
  isLearned: boolean;
  createdAt: string;
}

export type TermInputDifficulty = typeof TermInputDifficulty[keyof typeof TermInputDifficulty];


export const TermInputDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface TermInput {
  /** @minLength 1 */
  term: string;
  pronunciation?: string;
  /** @minLength 1 */
  definition: string;
  example?: string;
  /** @minLength 1 */
  category: string;
  difficulty: TermInputDifficulty;
}

export type TermUpdateDifficulty = typeof TermUpdateDifficulty[keyof typeof TermUpdateDifficulty];


export const TermUpdateDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export interface TermUpdate {
  /** @minLength 1 */
  term?: string;
  pronunciation?: string;
  /** @minLength 1 */
  definition?: string;
  example?: string;
  category?: string;
  difficulty?: TermUpdateDifficulty;
}

export interface QuizQuestion {
  termId: number;
  term: string;
  correctAnswer: string;
  choices: string[];
}

export interface CategoryCount {
  category: string;
  count: number;
  learned: number;
}

export interface DifficultyCount {
  difficulty: string;
  count: number;
}

export interface Stats {
  total: number;
  learned: number;
  favorites: number;
  byCategory: CategoryCount[];
  byDifficulty: DifficultyCount[];
}

export interface ErrorResponse {
  error: string;
}

export type ListTermsParams = {
/**
 * Search by term name or definition
 */
search?: string;
/**
 * Filter by category
 */
category?: string;
/**
 * Filter by difficulty
 */
difficulty?: ListTermsDifficulty;
/**
 * Return only favorited terms
 */
favoritesOnly?: boolean;
/**
 * Return only learned terms
 */
learnedOnly?: boolean;
};

export type ListTermsDifficulty = typeof ListTermsDifficulty[keyof typeof ListTermsDifficulty];


export const ListTermsDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;

export type GetQuizTermsParams = {
/**
 * Number of quiz terms to return
 */
count?: number;
/**
 * Filter quiz terms by category
 */
category?: string;
/**
 * Filter quiz terms by difficulty
 */
difficulty?: GetQuizTermsDifficulty;
};

export type GetQuizTermsDifficulty = typeof GetQuizTermsDifficulty[keyof typeof GetQuizTermsDifficulty];


export const GetQuizTermsDifficulty = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
} as const;