# lib/api-zod/src/generated/types/difficultyCount.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedTerms API - Medical Terms Dictionary and Flashcard App
 * OpenAPI spec version: 0.1.0
 */

export interface DifficultyCount {
  difficulty: string;
  count: number;
}