# lib/api-zod/src/generated/types/stats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedTerms API - Medical Terms Dictionary and Flashcard App
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { DifficultyCount } from './difficultyCount';

export interface Stats {
  total: number;
  learned: number;
  favorites: number;
  byCategory: CategoryCount[];
  byDifficulty: DifficultyCount[];
}