# lib/api-zod/src/generated/types/term.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedTerms API - Medical Terms Dictionary and Flashcard App
 * OpenAPI spec version: 0.1.0
 */
import type { TermDifficulty } from './termDifficulty';

export interface Term {
  id: number;
  term: string;
  /** @nullable */
  pronunciation?: string | null;
  definition: string;
  /** @nullable */
  example?: string | null;
  category: string;
  difficulty: TermDifficulty;
  isFavorite: boolean;
  isLearned: boolean;
  createdAt: Date;
}