# lib/api-zod/src/generated/types/termUpdate.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * MedTerms API - Medical Terms Dictionary and Flashcard App
 * OpenAPI spec version: 0.1.0
 */
import type { TermUpdateDifficulty } from './termUpdateDifficulty';

export interface TermUpdate {
  /** @minLength 1 */
  term?: string;
  pronunciation?: string;
  /** @minLength 1 */
  definition?: string;
  example?: string;
  category?: string;
  difficulty?: TermUpdateDifficulty;
}